- Moved getopt library to 3rd-party directory where it belongs  - fixed remaining compile errors, the toolset can be fully compiled using MSVC
git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@2566 21252fe5-60e8-0310-8ff0-c1ae1001594e

+++ src/squadt/processor.tcc
@@ -256,7 +256,7 @@ namespace squadt {
      
       /* Check whether outputs all exist and find the minimum timestamp of the inputs */
       BOOST_FOREACH(object_descriptor::sptr i, outputs) {
-        i->self_check(*g, maximum_input_timestamp);
+        i->self_check(*g, static_cast < const long int > (maximum_input_timestamp));
      
         result |= (i->status != object_descriptor::original) && (i->status != object_descriptor::reproducible_up_to_date);
       }
